--- libs/@blockprotocol/graph/src/types/subgraph/time.ts
@@ -1,5 +1,0 @@
-// Comment to capture intention of this file, this can hold types related to time and versioning. This could include
-// Timestamps, intervals, possibly different dimensions of time, depending on how the BP grows.
-
-/** @todo - Consider branding this */
-export type Timestamp = string;
