+++ libs/@blockprotocol/graph/src/types/temporal-versioning.ts
@@ -1,0 +1,5 @@
+/**
+ * Types used in embedding applications and blocks that support multi-axis temporal versioning schemes.
+ */
+
+export type Timestamp = string;
